ログイン済みの時に再度ログインしようとすると500エラーが起きる問題を修正

+++ src/main/java/jp/co/rakus/ec2018c/controller/LoginAfterController.java
@@ -60,8 +60,10 @@ public class LoginAfterController {
 			Order loginUserOrder = orderRepository.findByUserIdAndStatus(user.getId(), status);
 			
 			if(loginUserOrder != null) {
-				orderItemRepository.updateToOrderId(sessionOrder.getId(), loginUserOrder.getId());
-				orderRepository.deleteById(sessionOrder.getId());
+				if(sessionOrder != null && loginUserOrder != null) {
+					orderItemRepository.updateToOrderId(sessionOrder.getId(), loginUserOrder.getId());
+					orderRepository.deleteById(sessionOrder.getId());
+				}
 			}else {
 				sessionOrder.setUserId(user.getId());
 				orderRepository.save(sessionOrder);
